requirements.txt
amqp==2.6.1
asgiref==3.2.10
astroid==2.4.2
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
Django==3.1
django-appconf==1.0.4
django-celery-results==1.2.1
django-cryptography==1.0
django-debug-toolbar==2.2
django-denorm==1.0.0
django-redis==4.12.1
idna==2.10
isort==5.4.2
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.5
pycparser==2.20
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2020.1
redis==3.5.3
requests==2.24.0
rope==0.17.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
vine==1.3.0
wrapt==1.12.1
